+++ pheme/views.py
@@ -1,3 +1,5 @@
+import dataclasses
+
 from rest_framework.decorators import api_view, parser_classes
 from rest_framework.response import Response
 from pheme.parser.xml import XMLParser
@@ -22,9 +24,14 @@ def template(request):
         )
     )
     if input_flavour == 'gvmd':
-        if grouping == 'host':
-            grouping = scanreport.gvmd.group_by_host
         if grouping == 'nvt':
             grouping = scanreport.gvmd.group_by_nvt
-        return Response(scanreport.gvmd.transform(request.data, grouping))
+        else:
+            grouping = scanreport.gvmd.group_by_host
+
+        return Response(
+            dataclasses.asdict(
+                scanreport.gvmd.transform(request.data, grouping)
+            )
+        )
     return Response(request.data)