+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - XMLParser (pheme/parser/xml.py)
+- transformation for [GVMD] scan results to host grouped template data
+```
+curl -X POST\
+    'http://localhost:8000/template/?grouping=host'\
+     -H 'Content-Type: application/xml'\
+     -H 'Accept: application/json; indent=2'\
+     -d @path_to/scanreport.xml
+```
+- transformation for [GVMD] scan results to nvt grouped template data 
+```
+curl -X POST\
+    'http://localhost:8000/template/?grouping=nvt'\
+     -H 'Content-Type: application/xml'\
+     -H 'Accept: application/json; indent=2'\
+     -d @path_to/scanreport.xml
+```
 ### Changed
 ### Deprecated
 ### Removed
@@ -20,3 +36,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - django webserver
 - openapi (/openapi-schema/)
 - swagger (/docs/) 
+
+[GVMD]: https://github.com/greenbone/gvmd